requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2024.12.14
cloudinary==1.42.1
colorama==0.4.6
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
executing==2.2.0
gunicorn==23.0.0
icecream==2.1.4
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.10.1
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
